- Replacing our Config.php with a Robo\Config. - Catch bad config errors.
