+++ src/Application.php
@@ -7,10 +7,8 @@ use Aegir\Provision\Command\ServicesCommand;
 use Aegir\Provision\Command\ShellCommand;
 use Aegir\Provision\Command\StatusCommand;
 use Aegir\Provision\Command\VerifyCommand;
-use Aegir\Provision\Console\ActiveConfig;
 use Aegir\Provision\Console\Config;
 use Aegir\Provision\Console\ConsoleOutput;
-use Aegir\Provision\Console\DefaultsConfig;
 use Drupal\Console\Core\Style\DrupalStyle;
 use Psr\Log\LoggerInterface;
 use Psr\Log\LogLevel;
@@ -166,7 +164,7 @@ class Application extends BaseApplication
      */
     static function getAllContexts($name = '', $application = NULL) {
         $contexts = [];
-        $config = new ActiveConfig();
+        $config = new Config();
 
         $context_files = [];
         $finder = new \Symfony\Component\Finder\Finder();
